requirements.txt
python-whois
urllib
beautifulsoup4
requests
requests-html
html-to-etree
extract-social-media
fastapi
seolib
pyseoanalyzer
Twisted
Scrapy